# Optimize linear search of WAIT and WAITAOF when unblocking the client