--- a/src/server.h
+++ b/src/server.h
@@ -1267,6 +1267,7 @@ typedef struct client {
     sds peerid;                          /* Cached peer ID. */
     sds sockname;                        /* Cached connection target address. */
     listNode *client_list_node;          /* list node in client list */
+    listNode *client_waiting_acks_list_node;  /* list node in client waiting acks list. */
     listNode *postponed_list_node;       /* list node within the postponed list */
     void *module_blocked_client;         /* Pointer to the ValkeyModuleBlockedClient associated with this
                                           * client. This is set in case of module authentication before the