@enjoy-binbin 2024-07-15
An extreme test (wait 10 make sure it will get blocked, and the timeout is just 1ms so it can get unblocked in a short time)
```
./src/valkey-benchmark -c 5000 -n 1000000 wait 10 1
```

before:
```
Summary:
  throughput summary: 52452.14 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       85.334    27.792    80.575   153.215   166.911   190.591

Summary:
  throughput summary: 52645.43 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       84.970    27.888    80.383   154.751   164.991   185.215

Summary:
  throughput summary: 51894.13 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       86.001    29.632    80.767   153.215   163.583   193.279
```

after:
```
Summary:
  throughput summary: 61881.19 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       41.361    21.680    41.567    44.095    45.599    58.559

Summary:
  throughput summary: 62386.93 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       41.008    19.872    41.247    43.455    45.087    54.143

Summary:
  throughput summary: 61606.70 requests per second
  latency summary (msec):
          avg       min       p50       p95       p99       max
       41.515    19.424    41.695    43.839    45.631    54.943
```