@ranshid 2024-07-17
Thank you @enjoy-binbin. this optimization looks correct!

I would only suggest that we try and encapsulate all blocking related data in the BlockingState struct in the client.
Currently we keep the:
```
/* BLOCKED_WAIT and BLOCKED_WAITAOF */
    int numreplicas;      /* Number of replicas we are waiting for ACK. */
    int numlocal;         /* Indication if WAITAOF is waiting for local fsync. */
    long long reploffset; /* Replication offset to reach. */
```

inside the blocking state, maybe adding the list node ptr there is cleaner?