@enjoy-binbin 2024-08-04
> I guess it's OK but in the future, we can try to optimize the memory of the client struct. We could change this kind of lists to dicts or something else where we don't need to store a list node.

postponed_list_node also do the same thing, i think it is ok. yes, we can optimize it in the future.